feat: pss-cln-set-legacy-packages-002 - Fold testlink and testproviders into internal
