+++ docs/internal/baselines/ownership-inventory.json
@@ -4400,6 +4400,22 @@
       "destinationKind": "owner"
     },
     {
+      "packagePath": "pkg/services/operator_settings/internal/testlink",
+      "disposition": "move",
+      "destination": "operator_settings",
+      "destinationKind": "owner",
+      "successor": "pkg/services/operator_settings/internal",
+      "deletionCondition": "delete transitional top-level package after CLN-OPS-FOLD-TOPLEVEL cutover proof"
+    },
+    {
+      "packagePath": "pkg/services/operator_settings/internal/testproviders",
+      "disposition": "move",
+      "destination": "operator_settings",
+      "destinationKind": "owner",
+      "successor": "pkg/services/operator_settings/internal",
+      "deletionCondition": "delete transitional top-level package after CLN-OPS-FOLD-TOPLEVEL cutover proof"
+    },
+    {
       "packagePath": "pkg/services/operator_settings/servicewire",
       "disposition": "move",
       "destination": "operator_settings",
