+++ docs/internal/packaged-service-structure/package-target-manifest.json
@@ -343,6 +343,8 @@
     "pkg/services/operator_settings/internal/services/resolution",
     "pkg/services/operator_settings/internal/services/resolution/internal/service",
     "pkg/services/operator_settings/internal/services/resolution/wire",
+    "pkg/services/operator_settings/internal/testlink",
+    "pkg/services/operator_settings/internal/testproviders",
     "pkg/services/operator_settings/servicewire",
     "pkg/services/operator_settings/testlink",
     "pkg/services/operator_settings/testproviders",
@@ -2070,6 +2072,16 @@
       "packagePath": "pkg/services/operator_settings/internal/services/resolution/wire",
       "disposition": "retain",
       "destination": "operator_settings/internal/services/resolution"
+    },
+    {
+      "packagePath": "pkg/services/operator_settings/internal/testlink",
+      "disposition": "move",
+      "destination": "operator_settings/internal"
+    },
+    {
+      "packagePath": "pkg/services/operator_settings/internal/testproviders",
+      "disposition": "move",
+      "destination": "operator_settings/internal"
     },
     {
       "packagePath": "pkg/services/operator_settings/servicewire",
