+++ internal/ownershipinventory/operator_settings_mapping.go
@@ -23,6 +23,8 @@ var operatorSettingsMoveRules = []operatorSettingsMoveRule{
 	{exact: "testdata", prefix: "testdata/", subservice: ""},
 	{exact: "internal/service", prefix: "internal/service/", subservice: ""},
 	{exact: "internal/construct", prefix: "internal/construct/", subservice: ""},
+	{exact: "internal/testlink", prefix: "internal/testlink/", subservice: ""},
+	{exact: "internal/testproviders", prefix: "internal/testproviders/", subservice: ""},
 }
 
 func operatorSettingsMapping(packagePath string) (PackageRow, bool) {
