+++ pkg/services/operator_settings/internal/construct/testmain_test.go
@@ -3,10 +3,10 @@ package construct_test
 import (
 	"testing"
 
-	"github.com/portpowered/infinite-you/pkg/services/operator_settings/testlink"
+	internaltestlink "github.com/portpowered/infinite-you/pkg/services/operator_settings/internal/testlink"
 )
 
 func TestMain(m *testing.M) {
-	testlink.RegisterComposition()
+	internaltestlink.RegisterComposition()
 	m.Run()
 }
