+++ pkg/transports/cli/initsetup/testmain_test.go
@@ -3,10 +3,10 @@ package initsetup_test
 import (
 	"testing"
 
-	"github.com/portpowered/infinite-you/pkg/services/operator_settings/testlink"
+	internaltestlink "github.com/portpowered/infinite-you/pkg/services/operator_settings/internal/testlink"
 )
 
 func TestMain(m *testing.M) {
-	testlink.RegisterProvidersRoot()
+	internaltestlink.RegisterProvidersRoot()
 	m.Run()
 }
