+++ tests/functional/operator_settings/servicewire/testmain_test.go
@@ -3,11 +3,11 @@ package servicewire
 import (
 	"testing"
 
-	"github.com/portpowered/infinite-you/pkg/services/operator_settings/testlink"
+	internaltestlink "github.com/portpowered/infinite-you/pkg/services/operator_settings/internal/testlink"
 )
 
 // TestMain registers operator-settings composition hooks before servicewire tests run.
 func TestMain(m *testing.M) {
-	testlink.RegisterComposition()
+	internaltestlink.RegisterComposition()
 	m.Run()
 }
